+++ R/utility_functions.R
@@ -2436,3 +2436,12 @@ AssignNumberOfAsterisks <- function(value) {
   }
   return(label)
 }
+
+#' @keywords internal
+AdjustNumberOfCores <- function(num.of.cores) {
+  if (.Platform$OS.type == "windows" && num.of.cores > 1) {
+    message("On Windows, changing num.of.cores from ", num.of.cores, " to 1")
+    return(1)
+  }
+  return(num.of.cores)
+}